requirements.txt
allure-pytest==2.8.40
allure-python-commons==2.8.40
Appium-Python-Client==1.1.0
asgiref==3.3.4
attrs==21.2.0
base==1.0.4
blinker==1.4
Brotli==1.0.9
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.3.2
Faker==8.1.2
Flask==1.1.2
h11==0.12.0
h2==4.0.0
hpack==4.0.0
hyperframe==6.0.1
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jq==1.1.3
jsonpath==0.82
kaitaistruct==0.9
ldap3==2.8.1
MarkupSafe==1.1.1
mitmproxy==6.0.2
msgpack==1.0.2
packaging==20.9
passlib==1.7.4
pluggy==0.13.1
protobuf==3.14.0
publicsuffix2==2.20191221
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.2
pytest==6.2.4
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
selenium==3.141.0
six==1.16.0
sortedcontainers==2.3.0
text-unidecode==1.3
toml==0.10.2
tornado==6.1
urllib3==1.26.4
urwid==2.1.2
Werkzeug==1.0.1
wsproto==1.0.0
zstandard==0.14.1
